Adding dash app for viewing submissions and running the tests
I created a dash app based on the suggestion in the readme. Currently,
it lets one view the function definition for each submission and run
the pytest tests, outputting the results. I added some basic styling to
showcase that I understood css.

The app can be run with:
poetry run python app.py

One issue is that it does not display the helper function for problem 4.
To solve this without hard-coding, I would need to extend my ast parsing
to check for function calls to functions defined within the file.